requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
spacy>=3.1.0
nltk>=3.6.0
pytest>=7.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0 